Add auth middleware

+++ src/controllers/notes.ts
@@ -2,6 +2,7 @@ import { RequestHandler } from "express";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
 import NoteModel from "../models/note";
+import { assertIsDefined } from "../utils/assertIsDefined";
 
 interface NoteBody {
   title?: string;
@@ -13,8 +14,11 @@ interface UpdateNoteParams {
 }
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
   try {
-    const notes = await NoteModel.find().exec();
+    assertIsDefined(authenticatedUserId);
+
+    const notes = await NoteModel.find({ userId: authenticatedUserId }).exec();
     res.status(200).json(notes);
   } catch (error) {
     next(error);
@@ -22,9 +26,12 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 };
 
 export const getNoteById: RequestHandler = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
   const noteId = req.params.noteId;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid note id");
     }
@@ -33,6 +40,11 @@ export const getNoteById: RequestHandler = async (req, res, next) => {
     if (!note) {
       throw createHttpError(404, "Note not found");
     }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot access this note");
+    }
+
     res.status(200).json(note);
   } catch (error) {
     next(error);
@@ -47,12 +59,19 @@ export const createNote: RequestHandler<
 > = async (req, res, next) => {
   const title = req.body.title;
   const text = req.body.text;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!title) {
       throw createHttpError(400, "Note must have a title");
     }
-    const newNote = await NoteModel.create({ title, text });
+    const newNote = await NoteModel.create({
+      userId: authenticatedUserId,
+      title,
+      text,
+    });
     res.status(201).json(newNote);
   } catch (error) {
     next(error);
@@ -68,18 +87,28 @@ export const updateNote: RequestHandler<
   const noteId = req.params.noteId;
   const newTitle = req.body.title;
   const newText = req.body.text;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid note id");
     }
+
     if (!newTitle) {
       throw createHttpError(400, "Note must have a title");
     }
     const note = await NoteModel.findById(noteId).exec();
+
     if (!note) {
       throw createHttpError(404, "Note not found");
     }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot access this note");
+    }
+
     note.title = newTitle;
     note.text = newText;
     const updatedNote = await note.save();
@@ -91,14 +120,24 @@ export const updateNote: RequestHandler<
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authenticatedUserId = req.session.userId;
+
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid note id");
     }
     const note = await NoteModel.findById(noteId).exec();
+
     if (!note) {
       throw createHttpError(404, "Note not found");
     }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot access this note");
+    }
+
     await NoteModel.findByIdAndDelete(noteId);
     res.sendStatus(204);
   } catch (error) {
